requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.5
djangorestframework==3.16.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
flower==2.0.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
kombu==5.5.4
magic-filter==1.0.12
multidict==6.6.4
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
